Extract MFM normalize function
